Extract common package
